% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_plotfunctions.R
\name{heatmapClass}
\alias{heatmapClass}
\title{Heatmap of the similarities between classes about clustering}
\usage{
heatmapClass(output, ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompCluster}}

\item{...}{arguments to be passed to plot_ly}
}
\description{
Heatmap of the similarities between classes about clustering
}
\details{
The similarities between classes k and g is defined by 1 - Sigma(k,g)
Sigma(k,g)^2 = (1/n) * sum_{i=1}^n (P(Z_i=k|x_i) - P(Z_i=g|x_i))^2
}
\examples{
\donttest{
library(RMixtComp)

data(simData)
 
# define the algorithm's parameters
algo <- createAlgo()

# run RMixtCompt in unsupervised clustering mode + data as matrix
res <- mixtCompLearn(simDataLearn$matrix, desc$unsupervised, algo, nbClass = 2:4)
# plot
heatmapClass(res)

} 

}
\author{
Matthieu MARBAC
}
