% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_plotfunctions.R
\name{heatmapTikSorted}
\alias{heatmapTikSorted}
\title{Heatmap of the tik = P(Z_i=k|x_i)}
\usage{
heatmapTikSorted(output)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompCluster}}
}
\description{
Heatmap of the tik = P(Z_i=k|x_i)
}
\details{
Observation are sorted according to the hard partition then for each component
they are sorted by decreasing order of their tik's
}
\examples{
\dontrun{

library(RMixtComp)

# path to files
pathToData <- system.file("extdata", "data.csv", package = "RMixtComp")
pathToDescriptor <- system.file("extdata", "descUnsupervised.csv", package = "RMixtComp")

resGetData <- getData(c(pathToData, pathToDescriptor))


# define the algorithm's parameters
mcStrategy <- createMcStrategy()

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

# plot
heatmapTikSorted(res)

}

}
\author{
Matthieu MARBAC
}
