% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_plotfunctions.R
\name{heatmapVbles}
\alias{heatmapVbles}
\title{Heatmap of the similarities between variables about clustering}
\usage{
heatmapVbles(output)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompCluster}}
}
\description{
Heatmap of the similarities between variables about clustering
}
\details{
The similarities between variables j and h is defined by 1 - Delta(j,h)
Delta(j,h)^2 = (1/n) * sum_{i=1}^n sum_{k=1}^K (P(Z_i=k|x_{ij}) - P(Z_i=k|x_{ih}))^2
}
\examples{
\dontrun{

library(RMixtComp)

# path to files
pathToData <- system.file("extdata", "data.csv", package = "RMixtComp")
pathToDescriptor <- system.file("extdata", "descUnsupervised.csv", package = "RMixtComp")

resGetData <- getData(c(pathToData, pathToDescriptor))


# define the algorithm's parameters
mcStrategy <- list(nbBurnInIter = 100,
                   nbIter = 100,
                   nbGibbsBurnInIter = 50,
                   nbGibbsIter = 50,
                   ratioInitialization = 0.1)

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

# plot
heatmapVbles(res)

}

}
\author{
Matthieu MARBAC
}
