% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_plotfunctions.R
\name{histMisclassif}
\alias{histMisclassif}
\title{Histogram of the misclassification probabilities}
\usage{
histMisclassif(output, ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompCluster}}

\item{...}{arguments to be passed to plot_ly}
}
\description{
Histogram of the misclassification probabilities
}
\details{
Missclassification probability of observation i is denoted err_i
err_i = 1 - max_{k={1,\ldots,K}} P(Z_i=k|x_i)
Histgrams of err_i's can be plot for a specific class, all classes or every class
}
\examples{
\dontrun{

library(RMixtComp)

# path to files
pathToData <- system.file("extdata", "data.csv", package = "RMixtComp")
pathToDescriptor <- system.file("extdata", "descUnsupervised.csv", package = "RMixtComp")

resGetData <- getData(c(pathToData, pathToDescriptor))


# define the algorithm's parameters
mcStrategy <- createMcStrategy()

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

# plot
histMisclassif(res)

}

}
\author{
Matthieu MARBAC
}
