% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASSICCC_plotClasses.R
\name{plotComponentClusVis}
\alias{plotComponentClusVis}
\alias{plotObservationsClusVis}
\title{ClusVis plots}
\usage{
plotComponentClusVis(visInfo)
plotObservationsClusVis(visInfo)
}
\arguments{
\item{visInfo}{list. The parameters estimated by the clusvisMixtcomp function.}
}
\description{
plotComponentClusVis is a wrapper around the plotDensityClusVisu function of the ClusVisu package. It generates
a component interpretation graph using the visualization information returned by the clusvisMixtcomp function.
plotObservationsClusVis is a wrapper around the plotDensityClusVisu function of the ClusVisu package. It generates
a scatter-plot of the observation memberships using the visualization information returned by the clusvisMixtcomp function.
}
\examples{
\donttest{
library(RMixtComp)

data(simData)
 
# define the algorithm's parameters
algo <- createAlgo()

# run RMixtCompt in unsupervised clustering mode + data as matrix
res <- mixtCompLearn(simDataLearn$matrix, desc$unsupervised, algo, nbClass = 2:4)
resvisu <- clusvisMixtComp(res)

plotComponentClusVis(resvisu)
plotObservationsClusVis(resvisu) 

}

}
