% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_plotConvergence.R
\name{plotConvergence}
\alias{plotConvergence}
\title{Convergence of algorithm}
\usage{
plotConvergence(output, ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompCluster}}

\item{...}{graphical parameters}
}
\description{
Plot the evolution of the completed loglikelihood during the SEM algorithm. The vertical line denotes the end of the burn-in phase.
}
\details{
This function can be used to check the convergence and choose the parameters nbBurnInIter and nbIter from mcStrategy.
}
\examples{
\donttest{
library(RMixtComp)

data(simData)
 
# define the algorithm's parameters
algo <- createAlgo()

# run RMixtCompt in unsupervised clustering mode + data as matrix
res <- mixtCompLearn(simDataLearn$matrix, desc$unsupervised, algo, nbClass = 2:4)
# plot
plotConvergence(res)

} 

}
\author{
Quentin Grimonprez
}
