% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_plotUnivariateBoxplots.R
\name{plotDataBoxplot}
\alias{plotDataBoxplot}
\title{Boxplot per class}
\usage{
plotDataBoxplot(output, var, grl = TRUE, ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompCluster}}

\item{var}{name of the variable}

\item{grl}{if TRUE plot the general distribution of the data}

\item{...}{other parameters (see \emph{Details})}
}
\description{
Boxplot per class
}
\details{
For functional data, three other parameters are available:
\describe{
 \item{add.obs}{if TRUE, observations are added to the plot. Default = FALSE.}
 \item{ylim}{ylim of the plot.}
 \item{xlim}{xlim of the plot.}
}
}
\examples{
\donttest{
library(RMixtComp)

data(simData)
 
# define the algorithm's parameters
algo <- createAlgo()

# run RMixtCompt in unsupervised clustering mode + data as matrix
res <- mixtCompLearn(simDataLearn$matrix, desc$unsupervised, algo, nbClass = 2:4)
# plot
plotDataBoxplot(res, "Gaussian1")

} 

}
\author{
Matthieu MARBAC
}
