% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_plotUnivariateDistributions.R
\name{plotDataCI}
\alias{plotDataCI}
\title{Mean and 95\%-level confidence intervals per class}
\usage{
plotDataCI(output, var, class = 1:output$mixture$nbCluster, grl = FALSE,
  ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompCluster}}

\item{var}{name of the variable}

\item{class}{class to plot}

\item{grl}{if TRUE plot the CI for the dataset and not only classes}

\item{...}{other parameters (see \emph{Details})}
}
\description{
Mean and 95\%-level confidence intervals per class
}
\details{
For functional data, three other parameters are available:
\describe{
 \item{add.obs}{if TRUE, observations are added to the plot. Default = FALSE.}
 \item{ylim}{ylim of the plot.}
 \item{xlim}{xlim of the plot.}
}
}
\examples{
\dontrun{

library(RMixtComp)

# path to files
pathToData <- system.file("extdata", "data.csv", package = "RMixtComp")
pathToDescriptor <- system.file("extdata", "descUnsupervised.csv", package = "RMixtComp")

resGetData <- getData(c(pathToData, pathToDescriptor))


# define the algorithm's parameters
mcStrategy <- list(nbBurnInIter = 100,
                   nbIter = 100,
                   nbGibbsBurnInIter = 50,
                   nbGibbsIter = 50,
                   ratioInitialization = 0.1)

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

# plot
plotDataCI(res, "gaussian1")

} 

}
\author{
Matthieu MARBAC
}
