% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_plotConvergence.R
\name{plotParamConvergence}
\alias{plotParamConvergence}
\title{Evolution of parameters}
\usage{
plotParamConvergence(output, var, ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompCluster}}

\item{var}{name of the variable}

\item{...}{graphical parameters}
}
\description{
Plot the evolution of estimated parameters after the burn-in phase.
}
\examples{
\donttest{
library(RMixtComp)

data(simData)
 
# define the algorithm's parameters
algo <- createAlgo()

# run RMixtCompt in unsupervised clustering mode + data as matrix
res <- mixtCompLearn(simDataLearn$matrix, desc$unsupervised, algo, nbClass = 2:4)
# plot
plotParamConvergence(res, "Gaussian1")
plotParamConvergence(res, "Poisson1")

} 

}
\author{
Quentin Grimonprez
}
