% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASSICCC_plotClasses.R
\name{clusvisMixtComp}
\alias{clusvisMixtComp}
\title{Run clusvis for MixtComp output}
\usage{
clusvisMixtComp(resmixtcomp, sample.size = 5000, maxit = 10^3,
  nbrandomInit = 12, nbcpu = 1, estimateLogTik = FALSE)
}
\arguments{
\item{resmixtcomp}{list. It is an object return by MixComp.}

\item{sample.size}{numeric. Number of probabilities of classification sampled for parameter inference.}

\item{maxit}{numeric. It limits the number of iterations for the Quasi-Newton algorithm (default 1000).}

\item{nbrandomInit}{numeric. It defines the number of random initialization of the Quasi-Newton algorithm.}

\item{nbcpu}{numeric. It specifies the number of CPU (only for linux, R package parallel must by loaded).}

\item{estimateLogTik}{if TRUE, estimate the logtik. If FALSE, use the logtik returned by MixtComp}
}
\description{
This function estimates the parameters used for visualization of model-based clustering performs with R package ClusVis 
To achieve the parameter infernece, it automatically samples probabilities of classification from the model parameters
}
\examples{
\dontrun{

library(RMixtComp)

# path to files
pathToData <- system.file("extdata", "data.csv", package = "RMixtComp")
pathToDescriptor <- system.file("extdata", "descUnsupervised.csv", package = "RMixtComp")

resGetData <- getData(c(pathToData, pathToDescriptor))


# define the algorithm's parameters
mcStrategy <- createMcStrategy()

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

resvisu <- clusvisMixtComp(res)
}

}
