% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASSICCC_plotClasses.R
\name{plotComponentClusVis}
\alias{plotComponentClusVis}
\alias{plotObservationsClusVis}
\title{ClusVis plots}
\usage{
plotComponentClusVis(visInfo)
plotObservationsClusVis(visInfo)
}
\arguments{
\item{visInfo}{list. The parameters estimated by the clusvisMixtcomp function.}
}
\description{
plotComponentClusVis is a wrapper around the plotDensityClusVisu function of the ClusVisu package. It generates
a component interpretation graph using the visualization information returned by the clusvisMixtcomp function.
plotObservationsClusVis is a wrapper around the plotDensityClusVisu function of the ClusVisu package. It generates
a scatter-plot of the observation memberships using the visualization information returned by the clusvisMixtcomp function.
}
\examples{
\dontrun{

library(RMixtComp)

# path to files
pathToData <- system.file("extdata", "data.csv", package = "RMixtComp")
pathToDescriptor <- system.file("extdata", "descUnsupervised.csv", package = "RMixtComp")

resGetData <- getData(c(pathToData, pathToDescriptor))


# define the algorithm's parameters
mcStrategy <- createMcStrategy()

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

resvisu <- clusvisMixtComp(res)

plotComponentClusVis(resvisu)
plotObservationsClusVis(resvisu) 

}

}
