% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_plotConvergence.R
\name{plotConvergence}
\alias{plotConvergence}
\title{Convergence of algorithm}
\usage{
plotConvergence(output, ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompCluster}}

\item{...}{graphical parameters}
}
\description{
Plot the evolution of the completed loglikelihood during the SEM algorithm. The vertical line denotes the end of the burn-in phase.
}
\details{
This function can be used to check the convergence and choose the parameters nbBurnInIter and nbIter from mcStrategy.
}
\examples{
\dontrun{

library(RMixtComp)

# path to files
pathToData <- system.file("extdata", "data.csv", package = "RMixtComp")
pathToDescriptor <- system.file("extdata", "descUnsupervised.csv", package = "RMixtComp")

resGetData <- getData(c(pathToData, pathToDescriptor))


# define the algorithm's parameters
mcStrategy <- createMcStrategy()

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

# plot
plotConvergence(res)

} 

}
\author{
Quentin Grimonprez
}
