% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_plotConvergence.R
\name{plotParamConvergence}
\alias{plotParamConvergence}
\title{Evolution of parameters}
\usage{
plotParamConvergence(output, var, ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompCluster}}

\item{var}{name of the variable}

\item{...}{graphical parameters}
}
\description{
Plot the evolution of estimated parameters after the burn-in phase.
}
\examples{
\dontrun{

library(RMixtComp)

# path to files
pathToData <- system.file("extdata", "data.csv", package = "RMixtComp")
pathToDescriptor <- system.file("extdata", "descUnsupervised.csv", package = "RMixtComp")

resGetData <- getData(c(pathToData, pathToDescriptor))


# define the algorithm's parameters
mcStrategy <- createMcStrategy()

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

# plot
plotParamConvergence(res, "Gaussian1")
plotParamConvergence(res, "Poisson1")

} 

}
\author{
Quentin Grimonprez
}
