% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMixtCompPlot-package.R
\docType{package}
\name{RMixtCompPlot-package}
\alias{RMixtCompPlot-package}
\title{RMixtCompPlot}
\description{
Mixture Composer is a project to build mixture models with heterogeneous data
sets. The conditional independence property enables composing all the
data sets in a single mixture model.
}
\details{
\tabular{ll}{
Package: \tab MixtCompPlot\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2018-06-07\cr
}
}
\examples{
\dontrun{

library(RMixtComp)

# path to files
pathToData <- system.file("extdata", "data.csv", package = "RMixtComp")
pathToDescriptor <- system.file("extdata", "descUnsupervised.csv", package = "RMixtComp")

resGetData <- getData(c(pathToData, pathToDescriptor))


# define the algorithm's parameters
mcStrategy <- list(nbBurnInIter = 100,
                   nbIter = 100,
                   nbGibbsBurnInIter = 50,
                   nbGibbsIter = 50,
                   ratioInitialization = 0.1)

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

# run RMixtCompt for predicting
resPred <- mixtCompPredict(resGetData$lm, res$variable$param, mcStrategy, nbClass = 2, 
                           confidenceLevel = 0.95)



# Plot like Figure 1 (left)
plotDiscrimVbles(res)

# Plot like Figure 1 (right)
plotDiscrimClass(res)

# Plot like Figure 2 (left)
heatmapVbles(res)

# Plot like Figure 2 (right)
heatmapClass(res)

# Plot like Figure 3 (left)
histMisclassif(res)

# Plot like Figure 3 (right)
heatmapTikSorted(res)

}


}
\author{
Matthieu MARBAC, Quentin GRIMONPREZ

Maintainer: Quentin GRIMONPREZ  <quentin.grimonprez@inria.fr>
}
\keyword{package}
